30.08.18

+++ test.py
@@ -104,28 +104,45 @@ def cluster_image(img, colorspace, clusters, func, name):
     points = utils.image_to_points(img)
 
     labels = func(points, clusters)
-    res = utils.visual(points, labels, img.shape, name, colorspace)
+    res, mask = utils.visual(points, labels, img.shape, name, colorspace)
 
-    return res, labels
+    return res, labels, mask
 
 
 def test_image(folder, dst, func, names, clusters = 3):
-    shutil.rmtree(dst)
+    if os.path.isdir(dst):
+        shutil.rmtree(dst)
+
     os.mkdir(dst)
+    pts = 'pts'
+    clustered_pts = 'clustered_pts'
+    mask_dir = 'mask'
+
+    os.mkdir(os.path.join(dst, pts))
+    os.mkdir(os.path.join(dst, clustered_pts))
+    os.mkdir(os.path.join(dst, mask_dir))
 
     for i in os.listdir(folder):
         print('Processing [{0}] image'.format(i))
         img = cv2.imread(os.path.join(folder, i))
-        img = cv2.resize(img, (150,150))
+        img = cv2.resize(img, (100,100))
         i = os.path.splitext(i)[0]
         for idx, f in enumerate(func):
             try:
                 print('\tStarted processing {0} clustering algorithm'.format(names[idx]))
-                res, labels = cluster_image(img, None, clusters, f, '')
+                res, labels, mask = cluster_image(img, None, clusters, f, '')
                 print('\t\tClustered. Prepearing visualization')
                 l = np.unique(labels)
                 cv2.imwrite( os.path.join(dst, '{0}_{1}_{2}_clusters.png'.format(i, names[idx], l.shape[0])), res )
                 hyp.plot(utils.image_to_points(img), '.', group=labels, save_path=os.path.join(dst, '{0}_{1}_scatter.png'.format(i, names[idx])), show=False)
+
+                hyp.plot(utils.image_to_points(img), '.',
+                         save_path=os.path.join(dst, pts, '{0}_{1}_{2}.png'.format(i, names[idx], l.shape[0])), show=False)
+
+                hyp.plot(utils.image_to_points(img), '.', group=labels,
+                         save_path=os.path.join(dst, clustered_pts, '{0}_{1}_{2}.png'.format(i, names[idx], l.shape[0])), show=False)
+
+                cv2.imwrite( os.path.join(dst, mask_dir, '{0}_{1}_{2}.png'.format(i, names[idx], l.shape[0])), mask )
             except:
                 with open( os.path.join(dst, '{0}_{1}.txt'.format(i, names[idx])), 'w+') as f:
                     f.write('Error while processing')
@@ -133,13 +150,13 @@ def test_image(folder, dst, func, names, clusters = 3):
 
 
 def single_cluster(dim, samples, std, clusters, cluster_func):
-    x, y = make_blobs(n_samples=samples, centers=clusters, n_features=dim, random_state=1, cluster_std=std)
+    x, y = make_blobs(n_samples=samples, centers=clusters, n_features=dim, random_state=0, cluster_std=std)
     _y = cluster_func(x, clusters)
 
     acc = sklearn.metrics.homogeneity_score(y,_y)
 
-    hyp.plot(x, '.', group=y)
-    hyp.plot(x, '.', group=_y)
+    hyp.plot(x, '.', group=y, save_path='or.png')
+    hyp.plot(x, '.', group=_y, save_path='grp.png')
 
     print('Accuracy {0:0.2f}'.format(acc))
